+++ framework/wazuh/tests/test_stats.py
@@ -162,28 +162,28 @@ def test_get_daemons_stats_ko(mock_readfp):
             assert response.code == 1104, f'Response code is not the same'
 
 
-@pytest.mark.parametrize('daemon', [
+@pytest.mark.parametrize('component', [
     'logcollector', 'test'
 ])
-@patch('wazuh.core.agent.Agent.getstats')
+@patch('wazuh.core.agent.Agent.get_stats')
 @patch('wazuh.stats.get_agents_info', return_value=['000', '001'])
-def test_get_daemon_stats_json(mock_agents_info, mock_getstats, daemon):
-    """Test `get_daemon_stats_json` function from agent module.
+def test_get_agents_component_stats_json(mock_agents_info, mock_getstats, component):
+    """Test `get_agents_component_stats_json` function from agent module.
 
     Parameters
     ----------
     daemon : string
         Name of the daemon to get stats from.
     """
-    response = get_daemon_stats_json(agent_list=['001'], daemon=daemon)
+    response = get_agents_component_stats_json(agent_list=['001'], component=component)
     assert isinstance(response, AffectedItemsWazuhResult), f'The result is not AffectedItemsWazuhResult type'
-    mock_getstats.assert_called_once_with(daemon=daemon)
+    mock_getstats.assert_called_once_with(component=component)
 
 
-@patch('wazuh.core.agent.Agent.getstats')
+@patch('wazuh.core.agent.Agent.get_stats')
 @patch('wazuh.stats.get_agents_info', return_value=['000', '001'])
-def test_get_daemon_stats_json_ko(mock_agents_info, mock_getstats):
-    """Test `get_daemon_stats_json` function from agent module."""
-    response = get_daemon_stats_json(agent_list=['003'], daemon='logcollector')
+def test_get_agents_component_stats_json_ko(mock_agents_info, mock_getstats):
+    """Test `get_agents_component_stats_json` function from agent module."""
+    response = get_agents_component_stats_json(agent_list=['003'], component='logcollector')
     assert isinstance(response, AffectedItemsWazuhResult), f'The result is not AffectedItemsWazuhResult type'
     assert response.render()['data']['failed_items'][0]['error']['code'] == 1701, 'Expected error code was not returned'